+++ Common/src/main/java/fuzs/slotcycler/client/core/ClientAbstractions.java
@@ -1,0 +1,6 @@
+package fuzs.slotcycler.client.core;
+
+public interface ClientAbstractions {
+
+    int getRaisedDistance();
+}