+++ Common/src/main/java/fuzs/slotcycler/client/core/ClientModServices.java
@@ -1,0 +1,7 @@
+package fuzs.slotcycler.client.core;
+
+import fuzs.puzzleslib.util.PuzzlesUtil;
+
+public final class ClientModServices {
+    public static final ClientAbstractions ABSTRACTIONS = PuzzlesUtil.loadServiceProvider(ClientAbstractions.class);
+}