+++ Forge/src/main/java/fuzs/slotcycler/client/core/ForgeClientAbstractions.java
@@ -1,0 +1,9 @@
+package fuzs.slotcycler.client.core;
+
+public class ForgeClientAbstractions implements ClientAbstractions {
+
+    @Override
+    public int getRaisedDistance() {
+        return 0;
+    }
+}