+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.0.2-1.19.2] - 2022-10-03
+### Added
+- Added built-in support for [Raised] on Fabric
+- Cycle slots can now be set to only show while the scrolling modifier key (`Alt` by default) is held
+
 ## [v4.0.1-1.19.2] - 2022-09-30
 ### Added
 - Added a config option for a custom offset from the screen bottom to allow for compat with mods such as [Raised]